Fully Implemented Like buttons
Heart like buttons turn to red on hover and stay red if user is logged in, saving that individual bathroom to user's "likedBathroom" array in firebase. If the user isn't logged in, then it will direct them to the login page.

Also made changes to button.css changing it to button.scss and made changes to variable names in login.jsx to make more sense like "likedReviews" -> "likedBathrooms"